extract locations to a data file
